pvm windows x64

+++ scilab/modules/pvm/sci_gateway/c/sci_pvm_error_mode.c
@@ -50,12 +50,12 @@ int intspvm_error_mode(char *fname,unsigned long fname_len)
       int m;
       if ( pvm_error == 1 ) 
 	  { 
-		m = strlen(stop);
+		m = (int)strlen(stop);
 		CreateVarFromPtr(Rhs+1,STRING_DATATYPE,&m,&un,(loc=stop,&loc));
       } 
 	  else 
 	  { 
-		m = strlen(cont);
+		m = (int)strlen(cont);
 		CreateVarFromPtr(Rhs+1,STRING_DATATYPE,&m,&un,(loc=cont,&loc));
       }
       LhsVar(1)=Rhs+1;
